// work/ai-trading/src/components/home-page/interface.ts
export enum BOT_STATUS {
  ALL = '',
  OPEN = 'OPEN',
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE', // connect broker but dont trade
  PROCESSING = 'PROCESSING',
  // suspended: dont trade
  EXPIRED = 'EXPIRED',
  DISCONNECTED = 'DISCONNECTED',
  NOT_ENOUGH_BALANCE = 'NOT_ENOUGH_BALANCE',
  // tab not connected: not connect to broker, dont trade
  CONNECTING = 'CONNECTING',
  NOT_CONNECT = 'NOT_CONNECT',
}
// need to connect broker when user has these status:  DISCONNECTED | NOT_CONNECT
// need to disconnect broker when user has these status:  ACTIVE | INACTIVE

export interface BACK_TEST {
  bot_name: string;
  from: string;
  to: string;
  banner: {
    boxs: {
      [key: string]: {
        id: number;
        title: string;
        value: number;
      };
    };
    chart: { label: string; value: number }[];
    information: {
      initial_balance: number;
      average_trade: number;
      daily_avg_profit: number;
      drawdown: number;
      monthly_avg_profit: number;
      of_trades: number;
      percent_drawdown: number;
      profit_fator: number;
      return_do_ratio: number;
      sharpe_ratio: number;
      winning_percent: number;
    };
  };
  column_chart: {
    monthly_pln: { month: string; name: string; value: number }[];
    trades_of_weekly: { month: string; name: string; value: number }[];
  };
  performance: {
    apr: string;
    aug: string;
    dec: string;
    feb: string;
    jan: string;
    jul: string;
    jun: string;
    key: string;
    mar: string;
    may: string;
    nov: string;
    oct: string;
    sep: string;
    year: string;
    ytd: string;
  }[];
  pie_chart: {
    long_short_ratio: { type: string; value: number }[];
    pln_buy: { type: string; value: number }[];
    pln_sell: { type: string; value: number }[];
  };
  transaction: {
    average_loss: number;
    average_win: number;
    gross_loss: number;
    gross_profit: number;
    largest_loss: number;
    largest_win: number;
    max_consec_loss: number;
    max_consec_win: number;
    of_cancelled_expired: number;
    of_losses: number;
    of_wins: number;
  };
}

export interface BOT {
  broker?: string;
  broker_account?: string;
  broker_server?: string;
  bought: number;
  real_bought: string;
  code: string;
  created_at: string;
  currency: string;
  description: string;
  expires_at: string | undefined | null;
  id: string;
  image_url: string;
  name: string;
  max_drawdown: string;
  max_drawdown_change_percent: string;
  pnl: string;
  order: string;
  owner_created: string;
  display_price?: string;
  price: string;
  status: string;
  type: string;
  updated_at: string;
  user_status: Omit<BOT_STATUS, BOT_STATUS.ALL>;
  work_based_on: string[];
  translation: Record<
    Language,
    {
      description: BOT['description'];
      work_based_on: BOT['work_based_on'];
    }
  >;

  balance: number;
  current_balance: number;
  pnl_day_current: number;
  total_trade: number;

  back_test: string;
}